@@ -1,4 +1,4 @@
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Dashboard from "./components/Dashboard";
 import Login from "./components/Login";
 import Navbar from "./components/Navbar";
@@ -7,18 +7,11 @@ import Register from "./components/Register";
 function App() {
   return (
     <BrowserRouter>
-      <Switch>
-        <Route exact path="/">
-          <Login/>
-        </Route>
-        <Route path="/register">
-          <Register/>
-        </Route>
-        <Route path="/dashboard">
-          <Navbar/>
-          <Dashboard/>
-        </Route>
-      </Switch>
+      <Routes>
+        <Route exact path="/" element={<Login />}></Route>
+        <Route path="/register" element={<Register />}></Route>
+        <Route path="/dashboard" element={<><Navbar /><Dashboard /></>}></Route>
+      </Routes>
     </BrowserRouter>
   );
 }
