@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 const Login = () => {
     const [username, setUsername] = useState('');
     const [password, setPassword] = useState('');
     const [msg, setMsg] = useState('');
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const Auth = async (e) => {
         e.preventDefault();
@@ -15,7 +15,7 @@ const Login = () => {
                 username: username,
                 password: password
             });
-            history.push("/dashboard");
+            navigate("/dashboard");
         } catch (error) {
             if (error.response) {
                 setMsg(error.response.data.msg);
